% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_check_draws_array.R
\name{check_draws_array}
\alias{check_draws_array}
\title{Check a draws array}
\usage{
check_draws_array(draws)
}
\arguments{
\item{draws}{A three-dimensional array of posterior draws.}
}
\value{
Invisibly returns \code{NULL}. Throws an error if \code{draws} is invalid.
}
\description{
Checks that \code{draws} is a three-dimensional array with dimensions
corresponding to iterations, chains, and parameters.
}
\keyword{internal}
