% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_plot_mcmc_graph.R
\name{edge_color_from_graph}
\alias{edge_color_from_graph}
\title{Edge colour matching Fig. 2(c): black intersection, grey difference}
\usage{
edge_color_from_graph(graph)
}
\description{
Edge colour matching Fig. 2(c): black intersection, grey difference
}
\keyword{internal}
