% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_plot_mcmc_graph.R
\name{edge_lty_from_graph}
\alias{edge_lty_from_graph}
\title{Line type: solid intersection, dashed difference}
\usage{
edge_lty_from_graph(graph)
}
\description{
Line type: solid intersection, dashed difference
}
\keyword{internal}
