% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_mcmc_graph_summary.R
\name{format_cluster_summary_row}
\alias{format_cluster_summary_row}
\title{Format connected-component membership as one summary row}
\usage{
format_cluster_summary_row(cluster_result, chain_names, graph_name)
}
\description{
Format connected-component membership as one summary row
}
\keyword{internal}
