% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_get_chain_pair.R
\name{get_chain_pairs}
\alias{get_chain_pairs}
\title{Get all chain pairs for one parameter}
\usage{
get_chain_pairs(draws, parameter)
}
\arguments{
\item{draws}{A three dimensional array of posterior draws with dimensions
iterations by chains by parameters.}

\item{parameter}{A single parameter name. The name must be present in the
third dimension of 'draws'. This argument can not be 'NULL'.}
}
\value{
A list of chain-pair objects. Each element contains the index of the
first chain, the index of the second chian and the corresponding draws for the
selected parameter.
}
\description{
Extracts all unique pairs of chains for a selected parameter from a
'three dimensional draws array. For each pair, the function returns the chain
indices and the corresponding posterior draws
}
\keyword{internal}
