% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_plot_mcmc_graph.R
\name{n_monitored_parameters}
\alias{n_monitored_parameters}
\title{Number of monitored dimensions stored on a combined graph}
\usage{
n_monitored_parameters(graph)
}
\description{
Number of monitored dimensions stored on a combined graph
}
\keyword{internal}
