% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_plot_mcmc_graph.R
\name{edge_lty_from_graph}
\alias{edge_lty_from_graph}
\title{Line type matching Fig. 2(c): solid intersection, dashed difference}
\usage{
edge_lty_from_graph(graph)
}
\description{
Line type matching Fig. 2(c): solid intersection, dashed difference
}
\keyword{internal}
