% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_pairwise_rhat_clusters.R
\name{pairwise_rhat_clusters}
\alias{pairwise_rhat_clusters}
\title{Summarize clusters in a pairwise R-hat graph}
\usage{
pairwise_rhat_clusters(graph)
}
\arguments{
\item{graph}{An 'igraph' object}
}
\value{
A list containing multi-chain clusters. the number of multi-chain
clusters, isolated chains, the number of isolated chains, all connected
components, the total number of connected components, and the component
membership vector.
}
\description{
Finds connected components in a pairwise R-hat graph adn separates them into
multi-chain clusters and isolated chains. In this diagnostic, single chain
components are counted as isolated chains rather than as clusters.
}
