% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_pairwise_rhat_graph.R
\name{pairwise_rhat_graph}
\alias{pairwise_rhat_graph}
\title{Convert a pairwise R-hat matrix to a graph}
\usage{
pairwise_rhat_graph(rhat_matrix, rho = 1.01)
}
\arguments{
\item{rhat_matrix}{A square numeric matrix of pairwise R-hat values.}

\item{rho}{Numeric threshold used to decide whether two chains are connected.}
}
\value{
An undirected 'igraph' object whose nodes represent chains and whose
edges represent chain pairs with pairwise R-hat value below the threshold.
}
\description{
Converts a pairwsie R-hat matrix into an undirected graph. Each chain is a
node and an edge is added between two chains where their pairwise R-hat valuep
is less than or equal to 'rho'.
}
\keyword{internal}
