% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_split_chain.R
\name{split_chain}
\alias{split_chain}
\title{Split a chain into two halves}
\usage{
split_chain(x)
}
\arguments{
\item{x}{A numberical vector of draws from one chain for one parameter.}
}
\value{
A list with two numberic vectors: 'first', the first half of the chain,
and the 'second', the second half of the chain.
}
\description{
Split a numeric vector of MCMC draws into two equal halves. If the number of
draws is odd, the final draw is discarded so that both halves have the same
length.
}
\keyword{internal}
