% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_split_rhat.R
\name{split_rhat}
\alias{split_rhat}
\title{Compute split R-hat from split chains}
\usage{
split_rhat(split_draws)
}
\arguments{
\item{split_draws}{A numeric matrix of split-chain draws. Rows correspond to
draws and columns correspond to split chains.

@returns A numeric value giving the split R-hat statistic.

@keywords internal}
}
\description{
Computes the split R-hat statistic from a matrix of split-chain draws. Each
column is treated as one split chain, and each tow corresponds to one draw
within the split chain. This function calculates how different are the split
chains from each other, compared with how variable each split chain is internally
}
